backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
pydantic-settings
